+++ src/discord/monitor/message-handler.preflight.types.ts
@@ -34,6 +34,7 @@ export type DiscordMessagePreflightContext = {
   data: DiscordMessageEvent;
   client: Client;
   message: DiscordMessageEvent["message"];
+  messageChannelId: string;
   author: User;
   sender: DiscordSenderIdentity;